fix(build): restore lib, binary and lib-test compilation after three outside commits (bd-7cujy)
Fourteen compile errors from 849399a1b, 1f7d8df78 and 184f45661. Each fix
uses an API that already exists.

Non-test code (the reason the lib and the ee binary did not build):
- core/resume_transcripts.rs session_headers: sqlmodel-core 0.5.0 has no
  Value::Integer; bind the page size as Value::BigInt, the same i64.

Test code only:
- core/why.rs (5) and core/subscribe_poll_tests.rs (4): DbConnection has
  no public parameterized execute, so these fixture writes use
  execute_raw(&format!(..)), the established test idiom. Every interpolated
  value is a fixed test id or timestamp.
- cli/mod.rs: the three ResumeReport initializers gain
  transcript_history: ResumeTranscriptHistory::default(), the value
  production uses when there is no history; Op::CreateDirAll is a struct
  variant, now { mode: 0o700 } as in index-repair staging.

Verified at e3305f71f (rch clean overlay): the base fails with exactly the
14 errors; with this change cargo build --bin ee compiles, the lib tests
compile, and clippy --all-targets -D warnings exits 0. The lib tests that
still fail were born red with their features and never ran before this
change (three why_native_rule tests fail on their untouched fixture; one
resume transcript unicode test), or are already owned (bd-5tl24).

+++ src/core/resume_transcripts.rs
@@ -131,7 +131,7 @@ fn session_headers(
                 &[
                     Value::Text(workspace_id.to_owned()),
                     Value::Text(after.clone()),
-                    Value::Integer(RESUME_STORAGE_PAGE_SIZE as i64),
+                    Value::BigInt(RESUME_STORAGE_PAGE_SIZE as i64),
                 ],
             )
             .map_err(|_| storage_error())?;
